(currently building is broken) Splitting algorithms into a separate file and adding a test node
