+++ zebROS_ws/src/tf_object_detection/CMakeLists.txt
@@ -141,7 +141,8 @@ include_directories(
 ## Declare a C++ executable
 ## With catkin_make all packages are built within a single CMake context
 ## The recommended prefix ensures that target names across packages don't collide
-add_executable(tf_object_screen_to_world src/screen_to_world.cpp)
+add_executable(tf_object_screen_to_world src/screen_to_world.cpp src/depth_algorithms.cpp)
+add_executable(test_depth_detection src/test_depth_detection.cpp src/depth_algorithms.cpp)
 
 ## Rename C++ executable without prefix
 ## The above recommended prefix causes long target names, the following renames the
@@ -152,11 +153,15 @@ add_executable(tf_object_screen_to_world src/screen_to_world.cpp)
 ## Add cmake target dependencies of the executable
 ## same as for the library above
 add_dependencies(tf_object_screen_to_world ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
+add_dependencies(test_depth_detection ${${PROJECT_NAME}_EXPORTED_TARGETS} ${catkin_EXPORTED_TARGETS})
 
 ## Specify libraries to link a library or executable target against
 target_link_libraries(tf_object_screen_to_world
  ${catkin_LIBRARIES}
- )
+)
+target_link_libraries(test_depth_detection
+ ${catkin_LIBRARIES}
+)
 
 #############
 ## Install ##
